Do not assign floating numbers of cores in pargenes/jointsearch

+++ scripts/jointsearch/launch_multiple_jointsearch.py
@@ -35,7 +35,7 @@ def generate_scheduler_commands_file(families_dir, starting_tree, strategy, node
       if (0 == nodes_per_core):
         writer.write(family + " 1 " + str(tree_size) + " ")
       else:
-        writer.write(family + " " + str(max(1, tree_size / nodes_per_core)) + " " + str(tree_size) + " ")
+        writer.write(family + " " + str(max(1, int(tree_size / nodes_per_core))) + " " + str(tree_size) + " ")
       writer.write("-g " + gene_tree + " ")
       writer.write("-s " + os.path.join(family_path, "speciesTree.newick") +  " ")
       writer.write("-a " + os.path.join(family_path, "alignment.msa") +  " ")
